requirements.txt
google-cloud-translate>=3.11
